+++ src/main/kotlin/kitchenpos/menu/tobe/domain/MenuNamePolicy.kt
@@ -1,5 +1,12 @@
 package kitchenpos.menu.tobe.domain
 
-interface MenuNamePolicy {
-    fun validate(name: String)
+import kitchenpos.common.annotation.DomainService
+import kitchenpos.common.domain.Profanities
+
+@DomainService
+class MenuNamePolicy(private val profanities: Profanities) {
+    fun validate(name: String) {
+        require(name.isNotBlank()) { "메뉴 이름은 필수값입니다." }
+        require(!profanities.contains(name)) { "메뉴 이름에 금지어가 포함되어 있습니다." }
+    }
 }
