--- src/main/kotlin/kitchenpos/menu/tobe/domain/MenuProductPriceChangedListener.kt
@@ -1,7 +1,0 @@
-package kitchenpos.menu.tobe.domain
-
-import kitchenpos.product.tobe.domain.ProductPriceChangedEvent
-
-interface MenuProductPriceChangedListener {
-    fun handle(event: ProductPriceChangedEvent)
-}
